requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
httpx
pydantic-settings
resend
